refactor: move state
