+++ code/spikes/publish.001/src/types.mts
@@ -2,4 +2,4 @@ export * from './Pkg.Content.Bundle/types.mjs';
 export * from './Pkg.Content.Log/types.mjs';
 export * from './Markdown.File/types.mjs';
 
-export * from './ui/state/types.mjs';
+export * from './ui.logic/state/types.mjs';
