+++ code/spikes/publish.001/src/ui.logic/state/State.mts
@@ -1,6 +1,7 @@
-import { t } from '../common.mjs';
 import { QueryString } from './QueryString.mjs';
 import { BundlePaths } from '../../Pkg.Content.Bundle/Paths.mjs';
+
+export { BundlePaths, QueryString };
 
 /**
  * UI State Manager
