+++ code/spikes/publish.001/src/ui/Root/Root.tsx
@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from 'react';
 
-import { Color, COLORS, css, t } from '../common.mjs';
+import { Color, COLORS, css, t, State, BundlePaths } from '../common.mjs';
 import { Fetch } from '../Fetch.Util.mjs';
 import { History } from '../History/index.mjs';
-import { State } from '../state/index.mjs';
 import { MarkdownUtil } from '../Markdown/index.mjs';
-import { BundlePaths } from '../../Pkg.Content.Bundle/Paths.mjs';
 
 export type ShowMarkdownComponent = 'editor' | 'outline';
 
@@ -27,7 +25,7 @@ export const App: React.FC<AppProps> = (props) => {
       /**
        * Load log (history)
        */
-      const publicLog = await Fetch.json<t.PublicLogSummary>(State.BundlePaths.data.log);
+      const publicLog = await Fetch.json<t.PublicLogSummary>(BundlePaths.data.log);
       setLog(publicLog);
 
       /**
