+++ code/spikes/publish.001/src/ui/common.mts
@@ -1,1 +1,2 @@
 export * from './common/index.mjs';
+export * from '../ui.logic/index.mjs';
